backend/requirements.txt
Django==4.2
djangorestframework==3.20.4
djongo==1.3.6
pymongo==4.7.1
gunicorn==20.1.0
python-dotenv==1.0.0
